requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
